requirements.txt
python-telegram-bot==20.7
openai==1.12.0
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
sqlalchemy==2.0.23
requests==2.31.0
